+++ donation-tracker/donation-tracker.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Donation Tracker</title>
-    <link rel="stylesheet" href="styles.css" />
+    <link rel="stylesheet" href="donation-tracker-styles.css" />
   </head>
   <body>
     <header>
@@ -105,6 +105,6 @@
     <footer>
       <p>&copy; 2024 Donation Tracker. All rights reserved.</p>
     </footer>
-    <script src="script.js"></script>
+    <script src="donation-tracker-script.js"></script>
   </body>
 </html>
