+++ index.html
@@ -14,7 +14,9 @@
             <a href="index.html">Home</a>
           </li>
           <li>
-            <a href="donation-tracker/index.html">Donation Tracker</a>
+            <a href="donation-tracker/donation-tracker.html"
+              >Donation Tracker</a
+            >
           </li>
           <li>
             <a href="volunteer_hours_tracker.html">Volunteer Hours Tracker</a>
